+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed issue where parent records were not getting deleted from solr if empty. (DR-2342)
+- Fixed issue where suppressed captures were not being pulled back from solr. (DR-2470)
 
 ### Updated
 - Added additional rights statements that release high resolution permalinks (DR-2396)